Refactor cardinality estimation tests related to join into its own test suite
